useEnrichedRuns hook
